+++ mimesis/builtins/it.py
@@ -5,9 +5,9 @@
 import string
 from typing import Optional
 
-from mimesis import locales
 from mimesis.builtins.base import BaseSpecProvider
 from mimesis.enums import Gender
+from mimesis.locales import Locale
 from mimesis.typing import Seed
 
 __all__ = ["ItalySpecProvider"]
@@ -18,7 +18,7 @@ class ItalySpecProvider(BaseSpecProvider):
 
     def __init__(self, seed: Optional[Seed] = None) -> None:
         """Initialize attributes."""
-        super().__init__(locale=locales.IT, seed=seed)
+        super().__init__(locale=Locale.IT, seed=seed)
         self._pull(self._datafile)
 
     class Meta: