+++ mimesis/builtins/uk.py
@@ -3,9 +3,9 @@
 """Specific data provider for Ukraine (uk)."""
 from typing import List, Optional
 
-from mimesis import locales
 from mimesis.builtins.base import BaseSpecProvider
 from mimesis.enums import Gender
+from mimesis.locales import Locale
 from mimesis.typing import Seed
 
 __all__ = ["UkraineSpecProvider"]
@@ -16,7 +16,7 @@ class UkraineSpecProvider(BaseSpecProvider):
 
     def __init__(self, seed: Optional[Seed] = None) -> None:
         """Initialize attributes."""
-        super().__init__(locale=locales.UK, seed=seed)
+        super().__init__(locale=Locale.UK, seed=seed)
         self._pull(self._datafile)
 
     class Meta: